* fix RealTime division * switch sequencermapper back to file-backed shm, but with less frequent msync
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@4946 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ base/RealTime.h
@@ -1,3 +1,4 @@
+// -*- c-basic-offset: 4 -*-
 
 /*
     Rosegarden-4
@@ -77,7 +78,7 @@ struct RealTime
         else return sec <= r.sec;
     }
 
-    RealTime operator/(double d) const;
+    RealTime operator/(long d) const;
 
     // Find the fractional difference between times
     //
